requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
google-generativeai>=0.5.2
joblib==1.3.2
python-multipart==0.0.9
Pillow==10.2.0
openai>=1.0.0
python-dotenv==1.0.1
